+++ src/Entity/User/UserTrait.php
@@ -82,7 +82,7 @@ trait UserTrait
     /**
      * @return int
      */
-    public function getId()
+    public function getId(): ?int
     {
         return $this->id;
     }